將修改user權限合併至user.js內（controller/route）
