+++ app.js
@@ -8,7 +8,6 @@ const loginRouter = require("./routes/login");
 const usersRouter = require("./routes/users");
 const couponsRouter = require("./routes/coupons");
 const productsRouter = require("./routes/products");
-const changeUserPermissionRouter = require("./routes/changeUserPermission");
 const uploadRouter = require("./routes/upload");
 
 const app = express();
@@ -33,7 +32,6 @@ app.use("/api/v1/admin", loginRouter);
 app.use("/api/v1/admin/users", usersRouter);
 app.use("/api/v1/admin/coupons", couponsRouter);
 app.use("/api/v1/admin/products", productsRouter);
-app.use("/api/v1/admin/changeUserPermission", changeUserPermissionRouter);
 app.use("/api/v1/upload/image", uploadRouter);
 
 //404
